Separate all functions to different files
